// src/shared/enums/RightMenuLinkKeysEnum.ts
export enum RightMenuLinkKeysEnum {
    TODO = 'TODO',
    REFERRED = 'REFERRED',
    ACCEPTED = 'ACCEPTED',
    AWAITING_PUBLICATION = 'AWAITING_PUBLICATION',
    PUBLISHED = 'PUBLISHED',
    RETURNED = 'RETURNED',
    ALL = 'ALL',
    RETURNED_FROM_LEGAL_FILE = 'RETURNED_FROM_LEGAL_FILE',
    SENT = 'SENT',
    GROUP = 'GROUP',
    PERSONAL = 'PERSONAL',
    BORHAN_PERSONAL = 'BORHAN_PERSONAL',
    ANSWERED = 'ANSWERED',
}